+++ pysat/constellations/__init__.py
@@ -8,3 +8,5 @@ __all__ = ['testing', 'testing_empty', 'single_test']
 
 for const in __all__:
     exec("from pysat.constellations import {:}".format(const))
+
+del const